rename code files for runtime import map
