+++ packages/nextjs/src/editing/codegen/import-map.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable quotes */
 import { expect } from 'chai';
-import { combineImportEntries } from './default-import-map';
+import { combineImportEntries } from './import-map';
 import { ImportEntry } from '@sitecore-content-sdk/core/codegen';
 
 describe('Import Map Utils', () => {
