+++ packages/nextjs/src/editing/codegen/index.ts
@@ -1,1 +1,1 @@
-export { defaultImportEntries, combineImportEntries } from './default-import-map';
+export { defaultImportEntries, combineImportEntries } from './import-map';
